Add a tab style layout for large pageindex lists
After consideration of ajax, cookie and other schemes, went with
adding index tab to page parameter in the url, allowing it to be
applied and passed without concern for what method is serving
the actual page, i.e. keyword(x) provides x to an index list in the
keyword page. Each consumer of the page parameter needs to
be aware that it may contain out of band data.

Signed-off-by: geekwright <richard@geekwright.com>