Fix bug in creating search_body that forever cached garbage
Output of things like the RecentChanges were available for searching, 
even though the containing page has no relationship to the search term, and 
the change may well be ancient history when the search is done, and thus no 
longer even on the page that will be rendered. As a solution, zap PageIndex, 
RecentChanges, TOC, PageSetTOC before rendering search_body.